pyproject.toml
[project]
name = "agentic-retrieval"
version = "0.0.15"
description = "Planning, research, and report generation."
authors = [{ name = "Lance Martin" }]
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.55",
    "langchain-community>=0.3.9",
    "langchain-openai>=0.3.7",
    "langchain-ollama==0.2.2",
    "langchain-anthropic>=0.3.9",
    "openai>=1.61.0",
    "langchain-groq>=0.2.4",
    "requests>=2.32.3",
    "langchain-deepseek>=0.1.2",
    "python-dotenv==1.0.1",
    "langgraph_supervisor",
    "pytest",
    "httpx>=0.24.0",
    "markdownify>=0.11.6"
]

[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["agentic_retrieval"]

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = ["E", "F", "I", "D", "D401", "T201", "UP"]
lint.ignore = ["UP006", "UP007", "UP035", "D417", "E501"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
strict = true
